requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
discord.py==2.4.0
feedparser==6.0.11
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
sgmllib3k==1.0.0
typing==3.7.4.3
yarl==1.18.3
